trusted-types/HTMLElement-generic.tentative.html

<!DOCTYPE html>
<head>
  <script src="/resources/testharness.js"></script>
  <script src="/resources/testharnessreport.js"></script>
  <script src="support/helper.sub.js"></script>
</head>
<body>
<script>
  // TrustedURL Assignments
  let URLTestCases = [
    [ 'a', 'href' ],
    [ 'area', 'href' ],
    [ 'base', 'href' ],
    [ 'frame', 'src' ],
    [ 'iframe', 'src' ],
    [ 'img', 'src' ],
    [ 'input', 'src' ],
    [ 'link', 'href' ],
    [ 'video', 'src' ],
    [ 'object', 'data' ],
    [ 'object', 'codeBase' ],
    [ 'source', 'src' ],
    [ 'track', 'src' ]
  ];

  URLTestCases.forEach(c => {
    async_test(t => {
      assert_element_accepts_trusted_url(window, t, c[0], c[1], RESULTS.URL);
    }, c[0] + "." + c[1] + " assigned via policy (successful URL transformation)");
  });

  // TrustedScriptURL Assignments
  let ScriptURLTestCases = [
    [ 'embed', 'src' ],
    [ 'script', 'src' ]
  ];

  ScriptURLTestCases.forEach(c => {
    async_test(t => {
      assert_element_accepts_trusted_script_url(window, t, c[0], c[1], RESULTS.SCRIPTURL);
    }, c[0] + "." + c[1] + " assigned via policy (successful ScriptURL transformation)");
  });

  // TrustedHTML Assignments
  let HTMLTestCases = [
    [ 'div', 'innerHTML' ],
    [ 'iframe', 'srcdoc' ]
  ];

  HTMLTestCases.forEach(c => {
    async_test(t => {
      assert_element_accepts_trusted_html(window, t, c[0], c[1], RESULTS.HTML);
    }, c[0] + "." + c[1] + " assigned via policy (successful HTML transformation)");

  // TrustedScript Assignments
  let ScriptTestCases = [
    [ 'script', 'text' ]
  ];

  ScriptTestCases.forEach(c => {
    test(t => {
      assert_accepts_trusted_script(c[0], c[1]);
    }, c[0] + "." + c[1] + " accepts TrustedScript");
  });
</script>
